feat(dsp): vectorize the render path for 0.7.0
Stereo channel pairs share one complex transform with exactly conjugate-
symmetric twiddles; contiguous per-stage twiddles, permutation-free entry and
AArch64 NEON kernels with identical-order scalar fallbacks cover the FFT,
sinc spans over a mirrored ring, frame energy, bin magnitudes, overlap-add and
input validation. Analysis windows fetch only new source frames; phase advance
and rotation come from double-precision series. Output is no longer
byte-identical to 0.6.0: identical, opposite or silent channels agree within
float rounding, three exact tests became bounds, and the benchmark gains two
Master Tempo player cases. Version 0.7.0 across the package.

Validation: standalone suite 15/15 static and shared with installed C/C++ consumers, ASan/UBSan 15/15, wide 306-render matrix 0 contract failures with no in-band pitch change above 0.5 cent and no tone residual worse by more than 3 dB against 0.6.0, best-of-three benchmark 2.68x to 3.34x faster with GCC 7.5 -O2 on AArch64 and 2.50x to 2.69x with Clang 21 -O3 on Apple Silicon; device timing and music listening not performed.
Hardware: not touched

+++ src/envelope.cpp
@@ -13,11 +13,11 @@ void Envelope::configure(double milliseconds) noexcept {
     cutoff_ = std::min(size_ / 4, static_cast<std::size_t>(sample_rate_ * (milliseconds * .001)));
 }
 
-void Envelope::analyze(const double *magnitudes) noexcept {
+void Envelope::analyze(const float *magnitudes) noexcept {
     double maximum = 1e-12;
-    for (std::size_t k = 0; k <= size_ / 2; ++k) maximum = std::max(maximum, magnitudes[k]);
+    for (std::size_t k = 0; k <= size_ / 2; ++k) maximum = std::max(maximum, double(magnitudes[k]));
     for (std::size_t k = 0; k <= size_ / 2; ++k) {
-        scratch_[k] = Complex(static_cast<float>(std::log(std::max(magnitudes[k], maximum * 1e-5))), 0);
+        scratch_[k] = Complex(static_cast<float>(std::log(std::max(double(magnitudes[k]), maximum * 1e-5))), 0);
         if (k && k < size_ / 2) scratch_[size_ - k] = scratch_[k];
     }
     fft_.transform(scratch_.data(), true);
